[ID-49] Exception 모듈화 (#24)
* refact: move exception package

and refact kid personality package

* fix: ai client exception
