+++ src/main/java/com/ureca/idle/idleaiclient/implementation/AiException.java
@@ -1,0 +1,11 @@
+package com.ureca.idle.idleaiclient.implementation;
+
+
+import com.ureca.idle.idleexception.BaseException;
+import com.ureca.idle.idleexception.ExceptionType;
+
+public class AiException extends BaseException {
+    public AiException(ExceptionType exceptionType) {
+        super(exceptionType);
+    }
+}
