--- src/main/java/com/ureca/idle/idleapi/idleoriginapi/common/exception/user/UserException.java
@@ -1,13 +1,0 @@
-package com.ureca.idle.idleapi.idleoriginapi.common.exception.user;
-
-import com.ureca.idle.idleapi.idleoriginapi.common.exception.config.BaseException;
-import com.ureca.idle.idleapi.idleoriginapi.common.exception.config.ExceptionType;
-
-
-
-public class UserException extends BaseException {
-
-    public UserException(ExceptionType exceptionType) {
-        super(exceptionType);
-    }
-}
