+++ src/main/java/com/ureca/idle/idleapi/idleoriginapi/implementation/user/UserException.java
@@ -1,0 +1,13 @@
+package com.ureca.idle.idleapi.idleoriginapi.implementation.user;
+
+import com.ureca.idle.idleexception.BaseException;
+import com.ureca.idle.idleexception.ExceptionType;
+
+
+
+public class UserException extends BaseException {
+
+    public UserException(ExceptionType exceptionType) {
+        super(exceptionType);
+    }
+}
