+++ src/main/java/com/ureca/idle/idleapi/idleoriginapi/implementation/user/UserExceptionType.java
@@ -1,7 +1,7 @@
-package com.ureca.idle.idleapi.idleoriginapi.common.exception.user;
+package com.ureca.idle.idleapi.idleoriginapi.implementation.user;
 
-import com.ureca.idle.idleapi.idleoriginapi.common.exception.config.ExceptionType;
-import com.ureca.idle.idleapi.idleoriginapi.common.exception.config.Status;
+import com.ureca.idle.idleexception.ExceptionType;
+import com.ureca.idle.idleexception.Status;
 import lombok.AllArgsConstructor;
