+++ src/main/java/com/ureca/idle/idleexception/ExceptionResponse.java
@@ -1,0 +1,6 @@
+package com.ureca.idle.idleexception;
+
+
+
+public record ExceptionResponse(String message) {}
+
